Write reduce function to keep score
first wrote this in a seperate way: new hash created first, then an each method over the array which will then check for included substring.

Later changed to more concise reduce method which itself creates the new hash. One of the problems was the result not being called as return value for the reduce function, which gave a lot of errors. Never forget anymore...